Replaced old pratt implementation, added extensive docs
