+++ src/pratt.rs
@@ -1,530 +1,467 @@
-//! Pratt parser for efficiently parsing operators while respecting
-//! operator precedence.
+//! Utilities for parsing expressions using
+//! [Pratt parsing](https://en.wikipedia.org/wiki/Operator-precedence_parser#Pratt_parsing).
 //!
-//! Pratt parsing is an algorithm that allows efficient parsing of
-//! expressions using recursion.
+//! *“Who am I? What is my purpose in life? Does it really, cosmically speaking, matter if I don’t get up and go to work?”*
 //!
-//! The [`Parser::pratt`] method creates a [`Pratt`] parser. See the
-//! method's documentation for an example of how it can be used.
+//! Pratt parsing is a powerful technique for defining and parsing operators of varying arity, precedence, and
+//! associativity. Unlike [precedence climbing](https://en.wikipedia.org/wiki/Operator-precedence_parser), which
+//! defines operator precedence by structurally composing parsers of decreasing precedence, Pratt parsing defines
+//! precedence through a numerical
+//! ['binding power'](https://matklad.github.io/2020/04/13/simple-but-powerful-pratt-parsing.html#From-Precedence-to-Binding-Power)
+//! that determines how strongly operators should bind to the operands around them.
+//!
+//! Pratt parsers are defined with the [`Parser::pratt`] method.
+//!
+//! When writing pratt parsers, it is necessary to first define an 'atomic' operand used by the parser for building up
+//! expressions. In most languages, atoms are simple, self-delimiting patterns such as numeric and string literals,
+//! identifiers, or parenthesised expressions. Once an atom has been defined, operators can also be defined that
+//! operate upon said atoms.
+//!
+//! # Fold functions
+//!
+//! Because operators bind atoms together, pratt parsers require you to specify, for each operator, a function that
+//! combines its operands together into a syntax tree. These functions are given as the last arguments of [`infix`],
+//! [`prefix`], and [`postfix`].
+//!
+//! Fold functions have several overloads, allowing you to make use of only the operands, the operands and the
+//! operators, and even additionally a [`Span`] that covers the entire operation. See the documentation for each
+//! function to see which fold signatures can be used.
+//!
+//! # Examples
+//!
+//! ```
+//! use chumsky::prelude::*;
+//! use chumsky::pratt::*;
+//! use chumsky::extra;
+//!
+//! enum Expr {
+//!     Add(Box<Self>, Box<Self>),
+//!     Sub(Box<Self>, Box<Self>),
+//!     Pow(Box<Self>, Box<Self>),
+//!     Neg(Box<Self>),
+//!     Factorial(Box<Self>),
+//!     Deref(Box<Self>),
+//!     Literal(i32),
+//! }
+//!
+//! impl std::fmt::Display for Expr {
+//!     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+//!         match self {
+//!             Self::Literal(literal) => write!(f, "{literal}"),
+//!             Self::Factorial(left) => write!(f, "({left}!)"),
+//!             Self::Deref(left) => write!(f, "(*{left})"),
+//!             Self::Neg(right) => write!(f, "(-{right})"),
+//!             Self::Add(left, right) => write!(f, "({left} + {right})"),
+//!             Self::Sub(left, right) => write!(f, "({left} - {right})"),
+//!             Self::Pow(left, right) => write!(f, "({left} ^ {right})"),
+//!         }
+//!     }
+//! }
+//!
+//! let atom = text::int::<_, _, extra::Err<Simple<char>>>(10)
+//!     .from_str()
+//!     .unwrapped()
+//!     .map(Expr::Literal)
+//!     .padded();
+//!
+//! let op = |c| just(c).padded();
+//!
+//! let expr = atom.pratt((
+//!     // We want factorial to happen before any negation, so we need its precedence to be higher than `Expr::Neg`.
+//!     postfix(4, op('!'), |lhs| Expr::Factorial(Box::new(lhs))),
+//!     // Just like in math, we want that if we write -x^2, our parser parses that as -(x^2), so we need it to have
+//!     // exponents bind tighter than our prefix operators
+//!     infix(right(3), op('^'), |l, r| Expr::Pow(Box::new(l), Box::new(r))),
+//!     // Notice the conflict with our `Expr::Sub`. This will still parse correctly. We want negation to happen before
+//!     // `+` and `-`, so we set it's precedence higher.
+//!     prefix(2, op('-'), |rhs| Expr::Neg(Box::new(rhs))),
+//!     prefix(2, op('*'), |rhs| Expr::Deref(Box::new(rhs))),
+//!     // Our `-` and `+` bind the weakest, meaning that even if they occur first in an expression, they will be the
+//!     // last executed
+//!     infix(left(1), op('+'), |l, r| Expr::Add(Box::new(l), Box::new(r))),
+//!     infix(left(1), op('-'), |l, r| Expr::Sub(Box::new(l), Box::new(r))),
+//! ))
+//!     .map(|x| x.to_string());
+//!
+//! assert_eq!(
+//!     expr.parse("*1 + -2! - -3^2").into_result(),
+//!     Ok("(((*1) + (-(2!))) - (-(3 ^ 2)))".to_string()),
+//! );
+//! ```
 
-mod ops;
-use ops::Strength;
-pub use ops::{InfixOp, PostfixOp, PrefixOp};
+use super::*;
 
-use core::cmp::{self, Ordering};
+trait Operator<'a, I, O, E>
+where
+    I: Input<'a>,
+    E: ParserExtra<'a, I>,
+{
+    type Op;
+    type OpParser: Parser<'a, I, Self::Op, E>;
+    const IS_INFIX: bool = false;
+    const IS_PREFIX: bool = false;
+    const IS_POSTFIX: bool = false;
 
-use crate::{
-    extra::ParserExtra,
-    input::InputRef,
-    prelude::Input,
-    private::{Check, Emit, Mode, PResult, ParserSealed},
-    EmptyPhantom, Parser,
-};
-
-pub(super) use ops::{Infix, InfixPostfix, InfixPrefix, InfixPrefixPostfix, PrattOpOutput};
-
-/// Shorthand for [`InfixOp::new_left`].
-///
-/// Creates a left associative infix operator that is parsed with the
-/// parser `P`, and a function which is used to `build` a value `E`.
-/// The operator's precedence is determined by `strength`. The higher
-/// the value, the higher the precedence.
-pub fn left_infix<P, E, PO>(parser: P, strength: u8, build: InfixBuilder<E>) -> InfixOp<P, E, PO> {
-    InfixOp::new_left(parser, strength, build)
+    fn op_parser(&self) -> &Self::OpParser;
+    fn associativity(&self) -> Associativity;
+    fn fold_infix(&self, _lhs: O, _op: Self::Op, _rhs: O, _span: I::Span) -> O {
+        unreachable!()
+    }
+    fn fold_prefix(&self, _op: Self::Op, _rhs: O, _span: I::Span) -> O {
+        unreachable!()
+    }
+    fn fold_postfix(&self, _lhs: O, _op: Self::Op, _span: I::Span) -> O {
+        unreachable!()
+    }
 }
 
-/// Shorthand for [`InfixOp::new_right`].
+/// Defines the [associativity](https://en.wikipedia.org/wiki/Associative_property) and binding power of an [`infix`]
+/// operator (see [`left`] and [`right`]).
 ///
-/// Creates a right associative infix operator that is parsed with the
-/// parser `P`, and a function which is used to `build` a value `E`.
-/// The operator's precedence is determined by `strength`. The higher
-/// the value, the higher the precedence.
-pub fn right_infix<P, E, PO>(parser: P, strength: u8, build: InfixBuilder<E>) -> InfixOp<P, E, PO> {
-    InfixOp::new_right(parser, strength, build)
+/// Higher binding powers should be used for higher precedence operators.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum Associativity {
+    /// Specifies that the operator should be left-associative, with the given binding power (see [`left`]).
+    Left(u16),
+    /// Specifies that the operator should be right-associative, with the given binding power (see [`right`]).
+    Right(u16),
 }
 
-/// Shorthand for [`PrefixOp::new`].
+/// Specifies a left [`Associativity`] with the given binding power.
 ///
-/// Creates a prefix operator (a right-associative unary operator)
-/// that is parsed with the parser `P`, and a function which is used
-/// to `build` a value `E`. The operator's precedence is determined
-/// by `strength`. The higher the value, the higher the precedence.
-pub fn prefix<P, E, PO>(parser: P, strength: u8, build: PrefixBuilder<E>) -> PrefixOp<P, E, PO> {
-    PrefixOp::new(parser, strength, build)
+/// Left-associative operators are evaluated from the left-most terms, moving rightward. For example, the expression
+/// `a + b + c + d` will be evaluated as `((a + b) + c) + d` because addition is conventionally left-associative.
+pub fn left(binding_power: u16) -> Associativity {
+    Associativity::Left(binding_power)
 }
 
-/// Shorthand for [`PostfixOp::new`].
+/// Specifies a right [`Associativity`] with the given binding power.
 ///
-/// Creates a postfix operator (a left-associative unary operator)
-/// that is parsed with the parser `P`, and a function which is used
-/// to `build` a value `E`. The operator's precedence is determined
-/// by `strength`. The higher the value, the higher the precedence.
-pub fn postfix<P, E, PO>(parser: P, strength: u8, build: PostfixBuilder<E>) -> PostfixOp<P, E, PO> {
-    PostfixOp::new(parser, strength, build)
+/// Right-associative operators are evaluated from the right-most terms, moving leftward. For example, the expression
+/// `a ^ b ^ c ^ d` will be evaluated as `a ^ (b ^ (c ^ d))` because exponents are conventionally right-associative.
+pub fn right(binding_power: u16) -> Associativity {
+    Associativity::Right(binding_power)
 }
 
-/// A struct which represents a parser capable of using pratt-parsing.
+impl Associativity {
+    fn left_power(&self) -> u32 {
+        match self {
+            Self::Left(x) => *x as u32 * 2,
+            Self::Right(x) => *x as u32 * 2 + 1,
+        }
+    }
+
+    fn right_power(&self) -> u32 {
+        match self {
+            Self::Left(x) => *x as u32 * 2 + 1,
+            Self::Right(x) => *x as u32 * 2,
+        }
+    }
+}
+
+/// See [`infix`].
+pub struct Infix<A, F, Op, Args> {
+    op_parser: A,
+    fold: F,
+    associativity: Associativity,
+    #[allow(dead_code)]
+    phantom: EmptyPhantom<(Op, Args)>,
+}
+
+/// Specify a binary infix operator for a pratt parser with the given associativity, binding power, and
+/// [fold function](crate::pratt#fold-functions).
 ///
-/// This parser contains a parser of type `Atom`, which parses expressions that
-/// are separated by a set of operators of parsed by a parser of type `Ops`.
-/// The operators may have varying precedence levels, as well as associativity.
-/// For those unfamiliar with operator precedence and/or associativity, it may
-/// be helpful to read [this documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Operator_Precedence)
+/// Operators like addition, subtraction, multiplication, division, remainder, exponentiation, etc. are infix binary
+/// operators in most languages.
 ///
-/// This struct offers two methods:
-/// * `with_prefix_ops`: Attaches prefix operators to the parser
-/// * `with_postfix_ops`: Attaches postfix operators to the parser
+/// See [`left`] and [`right`] for information about associativity.
 ///
-/// Once one of the methods has been used, it will become unavailable
-/// due to the use of the type-state pattern to prevent accidental
-/// resetting of the operators.
-/// See [`Parser::pratt`] for an example of how to use these methods.
-pub struct Pratt<I, O, E, Atom, Ops> {
+/// The fold function (the last argument) must have one of the following signatures:
+///
+/// ```ignore
+/// // Combine the left and right operands
+/// impl Fn(O, O) -> O
+/// // Combine the left operand, the operator itself, and the right operand
+/// impl Fn(O, Op, O) -> O
+/// // Combine the left operand, the operator itself, the right operand, and the span that covers the whole operation
+/// impl Fn(O, Op, O, I::Span) -> O
+/// ```
+pub const fn infix<A, F, Op, Args>(
+    associativity: Associativity,
+    op_parser: A,
+    fold: F,
+) -> Infix<A, F, Op, Args> {
+    Infix {
+        op_parser,
+        fold,
+        associativity,
+        phantom: EmptyPhantom::new(),
+    }
+}
+
+macro_rules! infix_op {
+    (|$f:ident : Fn($($Arg:ty),*) -> O, $lhs:ident, $op:ident, $rhs:ident, $span:ident| $invoke:expr) => {
+        impl<'a, I, O, E, A, F, Op> Operator<'a, I, O, E> for Infix<A, F, Op, ($($Arg,)*)>
+        where
+            I: Input<'a>,
+            E: ParserExtra<'a, I>,
+            A: Parser<'a, I, Op, E>,
+            F: Fn($($Arg),*) -> O,
+        {
+            type Op = Op;
+            type OpParser = A;
+            const IS_INFIX: bool = true;
+            fn op_parser(&self) -> &Self::OpParser { &self.op_parser }
+            fn associativity(&self) -> Associativity { self.associativity }
+            fn fold_infix(&self, $lhs: O, $op: Self::Op, $rhs: O, $span: I::Span) -> O { let $f = &self.fold; $invoke }
+        }
+    };
+}
+
+// Allow `|lhs, rhs| <expr>` to be used as a fold closure for infix operators
+infix_op!(|f: Fn(O, O) -> O, lhs, _op, rhs, _span| f(lhs, rhs));
+// Allow `|lhs, op, rhs| <expr>` to be used as a fold closure for infix operators
+infix_op!(|f: Fn(O, Op, O) -> O, lhs, op, rhs, _span| f(lhs, op, rhs));
+// Allow `|lhs, op, rhs, span| <expr>` to be used as a fold closure for infix operators
+infix_op!(|f: Fn(O, Op, O, I::Span) -> O, lhs, op, rhs, span| f(lhs, op, rhs, span));
+
+/// See [`prefix`].
+pub struct Prefix<A, F, Op, Args> {
+    op_parser: A,
+    fold: F,
+    binding_power: u16,
+    #[allow(dead_code)]
+    phantom: EmptyPhantom<(Op, Args)>,
+}
+
+/// Specify a unary prefix operator for a pratt parser with the given binding power and
+/// [fold function](crate::pratt#fold-functions).
+///
+/// Operators like negation, not, dereferencing, etc. are prefix unary operators in most languages.
+///
+/// The fold function (the last argument) must have one of the following signatures:
+///
+/// ```ignore
+/// // Transform the operand
+/// impl Fn(O) -> O
+/// // Combine the operator itself and the operand
+/// impl Fn(Op, O) -> O
+/// // Combine the operator itself, the operand, and the span that covers the whole operation
+/// impl Fn(Op, O, I::Span) -> O
+/// ```
+pub const fn prefix<A, F, Op, Args>(
+    binding_power: u16,
+    op_parser: A,
+    fold: F,
+) -> Prefix<A, F, Op, Args> {
+    Prefix {
+        op_parser,
+        fold,
+        binding_power,
+        phantom: EmptyPhantom::new(),
+    }
+}
+
+macro_rules! prefix_op {
+    (|$f:ident : Fn($($Arg:ty),*) -> O, $op:ident, $rhs:ident, $span:ident| $invoke:expr) => {
+        impl<'a, I, O, E, A, F, Op> Operator<'a, I, O, E> for Prefix<A, F, Op, ($($Arg,)*)>
+        where
+            I: Input<'a>,
+            E: ParserExtra<'a, I>,
+            A: Parser<'a, I, Op, E>,
+            F: Fn($($Arg),*) -> O,
+        {
+            type Op = Op;
+            type OpParser = A;
+            const IS_PREFIX: bool = true;
+            fn op_parser(&self) -> &Self::OpParser { &self.op_parser }
+            fn associativity(&self) -> Associativity { Associativity::Left(self.binding_power) }
+            fn fold_prefix(&self, $op: Self::Op, $rhs: O, $span: I::Span) -> O { let $f = &self.fold; $invoke }
+        }
+    };
+}
+
+// Allow `|rhs| <expr>` to be used as a fold closure for prefix operators
+prefix_op!(|f: Fn(O) -> O, _op, rhs, _span| f(rhs));
+// Allow `|op, rhs| <expr>` to be used as a fold closure for prefix operators
+prefix_op!(|f: Fn(Op, O) -> O, op, rhs, _span| f(op, rhs));
+// Allow `|op, rhs, span| <expr>` to be used as a fold closure for prefix operators
+prefix_op!(|f: Fn(Op, O, I::Span) -> O, op, rhs, span| f(op, rhs, span));
+
+/// See [`postfix`].
+pub struct Postfix<A, F, Op, Args> {
+    op_parser: A,
+    fold: F,
+    binding_power: u16,
+    #[allow(dead_code)]
+    phantom: EmptyPhantom<(Op, Args)>,
+}
+
+/// Specify a unary postfix operator for a pratt parser with the given binding power and
+/// [fold function](crate::pratt#fold-functions).
+///
+/// Operators like factorial, field access, function composition, etc. are postfix unary operators in most languages.
+///
+/// The fold function (the last argument) must have one of the following signatures:
+///
+/// ```ignore
+/// // Transform the operand
+/// impl Fn(O) -> O
+/// // Combine the operand and the operator itself
+/// impl Fn(O, Op) -> O
+/// // Combine the operand, the operator itself, and the span that covers the whole operation
+/// impl Fn(Op, O, I::Span) -> O
+/// ```
+pub const fn postfix<A, F, Op, Args>(
+    binding_power: u16,
+    op_parser: A,
+    fold: F,
+) -> Postfix<A, F, Op, Args> {
+    Postfix {
+        op_parser,
+        fold,
+        binding_power,
+        phantom: EmptyPhantom::new(),
+    }
+}
+
+macro_rules! postfix_op {
+    (|$f:ident : Fn($($Arg:ty),*) -> O, $lhs:ident, $op:ident, $span:ident| $invoke:expr) => {
+        impl<'a, I, O, E, A, F, Op> Operator<'a, I, O, E> for Postfix<A, F, Op, ($($Arg,)*)>
+        where
+            I: Input<'a>,
+            E: ParserExtra<'a, I>,
+            A: Parser<'a, I, Op, E>,
+            F: Fn($($Arg),*) -> O,
+        {
+            type Op = Op;
+            type OpParser = A;
+            const IS_POSTFIX: bool = true;
+            fn op_parser(&self) -> &Self::OpParser { &self.op_parser }
+            fn associativity(&self) -> Associativity { Associativity::Left(self.binding_power) }
+            fn fold_postfix(&self, $lhs: O, $op: Self::Op, $span: I::Span) -> O { let $f = &self.fold; $invoke }
+        }
+    };
+}
+
+// Allow `|lhs| <expr>` to be used as a fold closure for postfix operators
+postfix_op!(|f: Fn(O) -> O, lhs, _op, _span| f(lhs));
+// Allow `|lhs, op| <expr>` to be used as a fold closure for postfix operators
+postfix_op!(|f: Fn(O, Op) -> O, lhs, op, _span| f(lhs, op));
+// Allow `|lhs, op, span| <expr>` to be used as a fold closure for postfix operators
+postfix_op!(|f: Fn(O, Op, I::Span) -> O, lhs, op, span| f(lhs, op, span));
+
+/// See [`Parser::pratt`].
+pub struct Pratt<Atom, Ops> {
     pub(crate) atom: Atom,
     pub(crate) ops: Ops,
-    pub(crate) _phantom: EmptyPhantom<(I, O, E)>,
 }
 
-impl<I, O, E, Atom, Ops> Copy for Pratt<I, O, E, Atom, Ops>
-where
-    Atom: Copy,
-    Ops: Copy,
-{
-}
+macro_rules! impl_pratt_for_tuple {
+    () => {};
+    ($head:ident $($X:ident)*) => {
+        impl_pratt_for_tuple!($($X)*);
+        impl_pratt_for_tuple!(~ $head $($X)*);
+    };
+    (~ $($X:ident)+) => {
+        #[allow(unused_variables, non_snake_case)]
+        impl<'a, Atom, $($X),*> Pratt<Atom, ($($X,)*)> {
+            fn pratt_go<M: Mode, I, O, E>(&self, inp: &mut InputRef<'a, '_, I, E>, min_power: u32) -> PResult<M, O>
+            where
+                I: Input<'a>,
+                E: ParserExtra<'a, I>,
+                Atom: Parser<'a, I, O, E>,
+                $($X: Operator<'a, I, O, E>),*
+            {
+                let pre_expr = inp.save();
+                let mut lhs = 'choice: {
+                    let ($($X,)*) = &self.ops;
 
-impl<I, O, E, Atom, Ops> Clone for Pratt<I, O, E, Atom, Ops>
-where
-    Atom: Clone,
-    Ops: Clone,
-{
-    fn clone(&self) -> Self {
-        Self {
-            atom: self.atom.clone(),
-            ops: self.ops.clone(),
-            _phantom: EmptyPhantom::new(),
-        }
-    }
-}
+                    // Prefix unary operators
+                    $(
+                        if $X::IS_PREFIX {
+                            match $X.op_parser().go::<M>(inp) {
+                                Ok(op) => {
+                                    match self.pratt_go::<M, _, _, _>(inp, $X.associativity().left_power()) {
+                                        Ok(rhs) => break 'choice M::combine(op, rhs, |op, rhs| {
+                                            let span = inp.span_since(pre_expr.offset());
+                                            $X.fold_prefix(op, rhs, span)
+                                        }),
+                                        Err(()) => inp.rewind(pre_expr),
+                                    }
+                                },
+                                Err(()) => inp.rewind(pre_expr),
+                            }
+                        }
+                    )*
 
-impl<'a, I, O, E, Atom, InfixOps, InfixOpsOut> Pratt<I, O, E, Atom, Infix<InfixOps, InfixOpsOut>> {
-    /// Extend a `Pratt` parser by setting prefix operators.
-    /// See [`Parser::pratt`] for an example of how to use this methods.
-    pub fn with_prefix_ops<PrefixOps, PrefixOpsOut>(
-        self,
-        prefix_ops: PrefixOps,
-    ) -> Pratt<I, O, E, Atom, InfixPrefix<InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut>>
-    where
-        I: Input<'a>,
-        E: ParserExtra<'a, I>,
-        InfixOps: Parser<'a, I, InfixOpsOut, E>,
-        PrefixOps: Parser<'a, I, PrefixOpsOut, E>,
-        Pratt<I, O, E, Atom, InfixPrefix<InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut>>:
-            PrattParser<'a, I, O, E>,
-    {
-        Pratt {
-            atom: self.atom,
-            ops: InfixPrefix {
-                infix: self.ops.infix,
-                prefix: prefix_ops,
-                _phantom: EmptyPhantom::new(),
-            },
-            _phantom: EmptyPhantom::new(),
-        }
-    }
+                    self.atom.go::<M>(inp)?
+                };
 
-    /// Extend a `Pratt` parser by setting postfix operators
-    /// See [`Parser::pratt`] for an example of how to use this method.
-    pub fn with_postfix_ops<PostfixOps, PostfixOpsOut>(
-        self,
-        postfix_ops: PostfixOps,
-    ) -> Pratt<I, O, E, Atom, InfixPostfix<InfixOps, InfixOpsOut, PostfixOps, PostfixOpsOut>>
-    where
-        I: Input<'a>,
-        E: ParserExtra<'a, I>,
-        InfixOps: Parser<'a, I, InfixOpsOut, E>,
-        PostfixOps: Parser<'a, I, PostfixOpsOut, E>,
-        Pratt<I, O, E, Atom, InfixPostfix<InfixOps, InfixOpsOut, PostfixOps, PostfixOpsOut>>:
-            PrattParser<'a, I, O, E>,
-    {
-        Pratt {
-            atom: self.atom,
-            ops: InfixPostfix {
-                infix: self.ops.infix,
-                postfix: postfix_ops,
-                _phantom: EmptyPhantom::new(),
-            },
-            _phantom: EmptyPhantom::new(),
-        }
-    }
-}
+                loop {
+                    let ($($X,)*) = &self.ops;
 
-impl<'a, I, O, E, Atom, InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut>
-    Pratt<I, O, E, Atom, InfixPrefix<InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut>>
-{
-    /// Extend a `Pratt` parser by setting postfix operators
-    pub fn with_postfix_ops<PostfixOps, PostfixOpsOut>(
-        self,
-        postfix_ops: PostfixOps,
-    ) -> Pratt<
-        I,
-        O,
-        E,
-        Atom,
-        InfixPrefixPostfix<
-            InfixOps,
-            InfixOpsOut,
-            PrefixOps,
-            PrefixOpsOut,
-            PostfixOps,
-            PostfixOpsOut,
-        >,
-    >
-    where
-        I: Input<'a>,
-        E: ParserExtra<'a, I>,
-        InfixOps: Parser<'a, I, InfixOpsOut, E>,
-        PrefixOps: Parser<'a, I, PrefixOpsOut, E>,
-        PostfixOps: Parser<'a, I, PostfixOpsOut, E>,
-        Pratt<
-            I,
-            O,
-            E,
-            Atom,
-            InfixPrefixPostfix<
-                InfixOps,
-                InfixOpsOut,
-                PrefixOps,
-                PrefixOpsOut,
-                PostfixOps,
-                PostfixOpsOut,
-            >,
-        >: PrattParser<'a, I, O, E>,
-    {
-        Pratt {
-            atom: self.atom,
-            ops: InfixPrefixPostfix {
-                infix: self.ops.infix,
-                prefix: self.ops.prefix,
-                postfix: postfix_ops,
-                _phantom: EmptyPhantom::new(),
-            },
-            _phantom: EmptyPhantom::new(),
-        }
-    }
-}
+                    let pre_op = inp.save();
 
-impl<'a, I, O, E, Atom, InfixOps, InfixOpsOut, PostfixOps, PostfixOpsOut>
-    Pratt<I, O, E, Atom, InfixPostfix<InfixOps, InfixOpsOut, PostfixOps, PostfixOpsOut>>
-{
-    /// Extend a `Pratt` parser by setting prefix operators
-    pub fn with_prefix_ops<PrefixOps, PrefixOpsOut>(
-        self,
-        prefix_ops: PrefixOps,
-    ) -> Pratt<
-        I,
-        O,
-        E,
-        Atom,
-        InfixPrefixPostfix<
-            InfixOps,
-            InfixOpsOut,
-            PrefixOps,
-            PrefixOpsOut,
-            PostfixOps,
-            PostfixOpsOut,
-        >,
-    >
-    where
-        I: Input<'a>,
-        E: ParserExtra<'a, I>,
-        PrefixOps: Parser<'a, I, PrefixOpsOut, E>,
-    {
-        Pratt {
-            atom: self.atom,
-            ops: InfixPrefixPostfix {
-                infix: self.ops.infix,
-                prefix: prefix_ops,
-                postfix: self.ops.postfix,
-                _phantom: EmptyPhantom::new(),
-            },
-            _phantom: EmptyPhantom::new(),
-        }
-    }
-}
+                    // Postfix unary operators
+                    $(
+                        let assoc = $X.associativity();
+                        if $X::IS_POSTFIX && assoc.right_power() >= min_power {
+                            match $X.op_parser().go::<M>(inp) {
+                                Ok(op) => {
+                                    lhs = M::combine(lhs, op, |lhs, op| {
+                                        let span = inp.span_since(pre_expr.offset());
+                                        $X.fold_postfix(lhs, op, span)
+                                    });
+                                    continue
+                                },
+                                Err(()) => inp.rewind(pre_op),
+                            }
+                        }
+                    )*
 
-type InfixBuilder<E> = fn(lhs: E, rhs: E) -> E;
+                    // Infix binary operators
+                    $(
+                        let assoc = $X.associativity();
+                        if $X::IS_INFIX && assoc.left_power() >= min_power {
+                            match $X.op_parser().go::<M>(inp) {
+                                Ok(op) => match self.pratt_go::<M, _, _, _>(inp, assoc.right_power()) {
+                                    Ok(rhs) => {
+                                        lhs = M::combine(
+                                            M::combine(lhs, rhs, |lhs, rhs| (lhs, rhs)),
+                                            op,
+                                            |(lhs, rhs), op| {
+                                                let span = inp.span_since(pre_expr.offset());
+                                                $X.fold_infix(lhs, op, rhs, span)
+                                            },
+                                        );
+                                        continue
+                                    },
+                                    Err(()) => inp.rewind(pre_op),
+                                },
+                                Err(()) => inp.rewind(pre_op),
+                            }
+                        }
+                    )*
 
-type PrefixBuilder<E> = fn(rhs: E) -> E;
-
-type PostfixBuilder<E> = fn(rhs: E) -> E;
-
-mod nameless_trait {
-    use super::*;
-
-    pub trait PrattParser<'a, I, Expr, E>
-    where
-        I: Input<'a>,
-        E: ParserExtra<'a, I>,
-    {
-        fn pratt_parse<M: Mode>(
-            &self,
-            inp: &mut InputRef<'a, '_, I, E>,
-            min_strength: Option<Strength>,
-        ) -> PResult<M, Expr>;
-    }
-}
-
-use nameless_trait::PrattParser;
-
-impl<'a, I, O, E, Atom, InfixOps, InfixOpsOut> PrattParser<'a, I, O, E>
-    for Pratt<I, O, E, Atom, Infix<InfixOps, InfixOpsOut>>
-where
-    I: Input<'a>,
-    E: ParserExtra<'a, I>,
-    Atom: Parser<'a, I, O, E>,
-    InfixOps: Parser<'a, I, PrattOpOutput<InfixBuilder<O>>, E>,
-{
-    fn pratt_parse<M>(
-        &self,
-        inp: &mut InputRef<'a, '_, I, E>,
-        min_strength: Option<Strength>,
-    ) -> PResult<M, O>
-    where
-        M: Mode,
-    {
-        let mut left = self.atom.go::<M>(inp)?;
-        loop {
-            let pre_op = inp.save();
-            let (op, prec) = match self.ops.infix.go::<Emit>(inp) {
-                Ok(PrattOpOutput(prec, build)) => {
-                    if prec.strength_left().is_lt(&min_strength) {
-                        inp.rewind(pre_op);
-                        return Ok(left);
-                    }
-                    (build, prec)
-                }
-                Err(_) => {
                     inp.rewind(pre_op);
-                    return Ok(left);
+                    break;
                 }
-            };
 
-            let right = self.pratt_parse::<M>(inp, Some(prec.strength_right()))?;
-            left = M::combine(left, right, op);
+                Ok(lhs)
+            }
         }
-    }
-}
 
-impl<'a, I, O, E, Atom, InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut> PrattParser<'a, I, O, E>
-    for Pratt<I, O, E, Atom, InfixPrefix<InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut>>
-where
-    I: Input<'a>,
-    E: ParserExtra<'a, I>,
-    Atom: Parser<'a, I, O, E>,
-    InfixOps: Parser<'a, I, PrattOpOutput<InfixBuilder<O>>, E>,
-    PrefixOps: Parser<'a, I, PrattOpOutput<PrefixBuilder<O>>, E>,
-{
-    fn pratt_parse<M>(
-        &self,
-        inp: &mut InputRef<'a, '_, I, E>,
-        min_strength: Option<Strength>,
-    ) -> PResult<M, O>
-    where
-        M: Mode,
-    {
-        let pre_op = inp.save();
-        let mut left = match self.ops.prefix.go::<Emit>(inp) {
-            Ok(PrattOpOutput(prec, build)) => {
-                let right = self.pratt_parse::<M>(inp, Some(prec.strength_right()))?;
-                M::map(right, build)
-            }
-            Err(_) => {
-                inp.rewind(pre_op);
-                self.atom.go::<M>(inp)?
-            }
-        };
-
-        loop {
-            let pre_op = inp.save();
-            let (op, prec) = match self.ops.infix.go::<Emit>(inp) {
-                Ok(PrattOpOutput(prec, build)) => {
-                    if prec.strength_left().is_lt(&min_strength) {
-                        inp.rewind(pre_op);
-                        return Ok(left);
-                    }
-                    (build, prec)
-                }
-                Err(_) => {
-                    inp.rewind(pre_op);
-                    return Ok(left);
-                }
-            };
-
-            let right = self.pratt_parse::<M>(inp, Some(prec.strength_right()))?;
-            left = M::combine(left, right, op);
-        }
-    }
-}
-
-impl<'a, I, O, E, Atom, InfixOps, InfixOpsOut, PostfixOps, PostfixOpsOut> PrattParser<'a, I, O, E>
-    for Pratt<I, O, E, Atom, InfixPostfix<InfixOps, InfixOpsOut, PostfixOps, PostfixOpsOut>>
-where
-    I: Input<'a>,
-    E: ParserExtra<'a, I>,
-    Atom: Parser<'a, I, O, E>,
-    InfixOps: Parser<'a, I, PrattOpOutput<InfixBuilder<O>>, E>,
-    PostfixOps: Parser<'a, I, PrattOpOutput<PostfixBuilder<O>>, E>,
-{
-    fn pratt_parse<M>(
-        &self,
-        inp: &mut InputRef<'a, '_, I, E>,
-        min_strength: Option<Strength>,
-    ) -> PResult<M, O>
-    where
-        M: Mode,
-    {
-        let mut left = self.atom.go::<M>(inp)?;
-        loop {
-            let pre_op = inp.save();
-            match self.ops.postfix.go::<Emit>(inp) {
-                Ok(PrattOpOutput(prec, build)) => {
-                    if prec.strength_left().is_lt(&min_strength) {
-                        inp.rewind(pre_op);
-                        return Ok(left);
-                    }
-                    left = M::map(left, build);
-                    continue;
-                }
-                Err(_) => {
-                    inp.rewind(pre_op);
-                }
-            }
-
-            let (op, prec) = match self.ops.infix.go::<Emit>(inp) {
-                Ok(PrattOpOutput(prec, build)) => {
-                    if prec.strength_left().is_lt(&min_strength) {
-                        inp.rewind(pre_op);
-                        return Ok(left);
-                    }
-                    (build, prec)
-                }
-                Err(_) => {
-                    inp.rewind(pre_op);
-                    return Ok(left);
-                }
-            };
-
-            let right = self.pratt_parse::<M>(inp, Some(prec.strength_right()))?;
-            left = M::combine(left, right, op);
-        }
-    }
-}
-
-impl<
-        'a,
-        I,
-        O,
-        E,
-        Atom,
-        InfixOps,
-        InfixOpsOut,
-        PrefixOps,
-        PrefixOpsOut,
-        PostfixOps,
-        PostfixOpsOut,
-    > PrattParser<'a, I, O, E>
-    for Pratt<
-        I,
-        O,
-        E,
-        Atom,
-        InfixPrefixPostfix<
-            InfixOps,
-            InfixOpsOut,
-            PrefixOps,
-            PrefixOpsOut,
-            PostfixOps,
-            PostfixOpsOut,
-        >,
-    >
-where
-    I: Input<'a>,
-    E: ParserExtra<'a, I>,
-    Atom: Parser<'a, I, O, E>,
-    InfixOps: Parser<'a, I, PrattOpOutput<InfixBuilder<O>>, E>,
-    PrefixOps: Parser<'a, I, PrattOpOutput<PrefixBuilder<O>>, E>,
-    PostfixOps: Parser<'a, I, PrattOpOutput<PostfixBuilder<O>>, E>,
-{
-    fn pratt_parse<M>(
-        &self,
-        inp: &mut InputRef<'a, '_, I, E>,
-        min_strength: Option<Strength>,
-    ) -> PResult<M, O>
-    where
-        M: Mode,
-    {
-        let pre_op = inp.save();
-        let mut left = match self.ops.prefix.go::<Emit>(inp) {
-            Ok(PrattOpOutput(prec, build)) => {
-                let right = self.pratt_parse::<M>(inp, Some(prec.strength_right()))?;
-                M::map(right, build)
-            }
-            Err(_) => {
-                inp.rewind(pre_op);
-                self.atom.go::<M>(inp)?
-            }
-        };
-
-        loop {
-            let pre_op = inp.save();
-            match self.ops.postfix.go::<Emit>(inp) {
-                Ok(PrattOpOutput(prec, build)) => {
-                    if prec.strength_left().is_lt(&min_strength) {
-                        inp.rewind(pre_op);
-                        return Ok(left);
-                    }
-                    left = M::map(left, build);
-                    continue;
-                }
-                Err(_) => {
-                    inp.rewind(pre_op);
-                }
-            }
-
-            let (op, prec) = match self.ops.infix.go::<Emit>(inp) {
-                Ok(PrattOpOutput(prec, build)) => {
-                    if prec.strength_left().is_lt(&min_strength) {
-                        inp.rewind(pre_op);
-                        return Ok(left);
-                    }
-                    (build, prec)
-                }
-                Err(_) => {
-                    inp.rewind(pre_op);
-                    return Ok(left);
-                }
-            };
-
-            let right = self.pratt_parse::<M>(inp, Some(prec.strength_right()))?;
-            left = M::combine(left, right, op);
-        }
-    }
-}
-
-macro_rules! impl_parse {
-    ($Parser:ident < $($Gen:ident),+ $(,)?>) => {
-        impl<'a, I, O, E, Atom, $($Gen),+> ParserSealed<'a, I, O, E> for Pratt<I, O, E, Atom, $Parser<$($Gen),+>>
+        #[allow(unused_variables, non_snake_case)]
+        impl<'a, I, O, E, Atom, $($X),*> ParserSealed<'a, I, O, E> for Pratt<Atom, ($($X,)*)>
         where
             I: Input<'a>,
             E: ParserExtra<'a, I>,
             Atom: Parser<'a, I, O, E>,
-            Self: PrattParser<'a, I, O, E>,
+            $($X: Operator<'a, I, O, E>),*
         {
-            fn go<M: Mode>(&self, inp: &mut InputRef<'a, '_, I, E>) -> PResult<M, O>
-            where
-                Self: Sized,
-            {
-                self.pratt_parse::<M>(inp, None)
+            fn go<M: Mode>(&self, inp: &mut InputRef<'a, '_, I, E>) -> PResult<M, O> {
+                self.pratt_go::<M, _, _, _>(inp, 0)
             }
 
             go_extra!(O);
@@ -532,23 +469,47 @@ macro_rules! impl_parse {
     };
 }
 
-impl_parse!(Infix<InfixOps, InfixOpsOut>);
-
-impl_parse!(InfixPrefix<InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut>);
-
-impl_parse!(InfixPostfix<InfixOps, InfixOpsOut, PostfixOps, PostfixOpsOut>);
-
-impl_parse!(InfixPrefixPostfix<InfixOps, InfixOpsOut, PrefixOps, PrefixOpsOut, PostfixOps, PostfixOpsOut,>);
+impl_pratt_for_tuple!(A_ B_ C_ D_ E_ F_ G_ H_ I_ J_ K_ L_ M_ N_ O_ P_ Q_ R_ S_ T_ U_ V_ W_ X_ Y_ Z_);
 
 #[cfg(test)]
 mod tests {
-    use crate::error::Error;
-    use crate::extra::Err;
-    use crate::prelude::{choice, end, just, Simple, SimpleSpan};
-    use crate::util::MaybeRef;
-    use crate::{text, ParseResult};
-
     use super::*;
+    use crate::{extra::Err, prelude::*};
+
+    fn factorial(x: i64) -> i64 {
+        if x == 0 {
+            1
+        } else {
+            x * factorial(x - 1)
+        }
+    }
+
+    fn parser<'a>() -> impl Parser<'a, &'a str, i64> {
+        let atom = text::int(10).padded().from_str::<i64>().unwrapped();
+
+        atom.pratt((
+            prefix(2, just('-'), |x: i64| -x),
+            postfix(2, just('!'), factorial),
+            infix(left(0), just('+'), |l, r| l + r),
+            infix(left(0), just('-'), |l, r| l - r),
+            infix(left(1), just('*'), |l, r| l * r),
+            infix(left(1), just('/'), |l, _, r| l / r),
+        ))
+    }
+
+    #[test]
+    fn precedence() {
+        assert_eq!(parser().parse("2 + 3 * 4").into_result(), Ok(14));
+        assert_eq!(parser().parse("2 * 3 + 4").into_result(), Ok(10));
+    }
+
+    #[test]
+    fn unary() {
+        assert_eq!(parser().parse("-2").into_result(), Ok(-2));
+        assert_eq!(parser().parse("4!").into_result(), Ok(24));
+        assert_eq!(parser().parse("2 + 4!").into_result(), Ok(26));
+        assert_eq!(parser().parse("-2 + 2").into_result(), Ok(0));
+    }
 
     enum Expr {
         Literal(i64),
@@ -580,21 +541,27 @@ mod tests {
         }
     }
 
-    fn parser<'a>() -> impl Parser<'a, &'a str, String, Err<Simple<'a, char>>> {
+    fn u(e: fn(Box<Expr>) -> Expr, r: Expr) -> Expr {
+        e(Box::new(r))
+    }
+    fn i(e: fn(Box<Expr>, Box<Expr>) -> Expr, l: Expr, r: Expr) -> Expr {
+        e(Box::new(l), Box::new(r))
+    }
+
+    fn expr_parser<'a>() -> impl Parser<'a, &'a str, String, Err<Simple<'a, char>>> {
         let atom = text::int(10).from_str().unwrapped().map(Expr::Literal);
 
-        let operator = choice((
-            left_infix(just('+'), 0, |l, r| Expr::Add(Box::new(l), Box::new(r))),
-            left_infix(just('-'), 0, |l, r| Expr::Sub(Box::new(l), Box::new(r))),
-            right_infix(just('*'), 1, |l, r| Expr::Mul(Box::new(l), Box::new(r))),
-            right_infix(just('/'), 1, |l, r| Expr::Div(Box::new(l), Box::new(r))),
-        ));
-
-        atom.pratt(operator).map(|x| x.to_string())
+        atom.pratt((
+            infix(left(0), just('+'), |l, r| i(Expr::Add, l, r)),
+            infix(left(0), just('-'), |l, r| i(Expr::Sub, l, r)),
+            infix(right(1), just('*'), |l, r| i(Expr::Mul, l, r)),
+            infix(right(1), just('/'), |l, r| i(Expr::Div, l, r)),
+        ))
+        .map(|x| x.to_string())
     }
 
     fn complete_parser<'a>() -> impl Parser<'a, &'a str, String, Err<Simple<'a, char>>> {
-        parser().then_ignore(end())
+        expr_parser().then_ignore(end())
     }
 
     fn parse(input: &str) -> ParseResult<String, Simple<char>> {
@@ -602,7 +569,7 @@ mod tests {
     }
 
     fn parse_partial(input: &str) -> ParseResult<String, Simple<char>> {
-        parser().lazy().parse(input)
+        expr_parser().lazy().parse(input)
     }
 
     fn unexpected<'a, C: Into<Option<MaybeRef<'a, char>>>, S: Into<SimpleSpan>>(
@@ -666,24 +633,22 @@ mod tests {
             .unwrapped()
             .map(Expr::Literal);
 
-        let operator = choice((
-            left_infix(just('+'), 0, |l, r| Expr::Add(Box::new(l), Box::new(r))),
-            left_infix(just('-'), 0, |l, r| Expr::Sub(Box::new(l), Box::new(r))),
-            right_infix(just('*'), 1, |l, r| Expr::Mul(Box::new(l), Box::new(r))),
-            right_infix(just('/'), 1, |l, r| Expr::Div(Box::new(l), Box::new(r))),
-        ));
-
         let parser = atom
-            .pratt(operator)
-            .with_prefix_ops(choice((
+            .pratt((
+                // -- Prefix
                 // Because we defined '*' and '/' as right associative operators,
                 // in order to get these to function as expected, their strength
                 // must be higher
-                prefix(just('-'), 2, |rhs| Expr::Negate(Box::new(rhs))),
-                prefix(just('~'), 2, |rhs| Expr::Not(Box::new(rhs))),
+                prefix(2, just('-'), |r| u(Expr::Negate, r)),
+                prefix(2, just('~'), |r| u(Expr::Not, r)),
                 // This is what happens when not
-                prefix(just('§'), 1, |rhs| Expr::Confusion(Box::new(rhs))),
-            )))
+                prefix(1, just('§'), |r| u(Expr::Confusion, r)),
+                // -- Infix
+                infix(left(0), just('+'), |l, r| i(Expr::Add, l, r)),
+                infix(left(0), just('-'), |l, r| i(Expr::Sub, l, r)),
+                infix(right(1), just('*'), |l, r| i(Expr::Mul, l, r)),
+                infix(right(1), just('/'), |l, r| i(Expr::Div, l, r)),
+            ))
             .map(|x| x.to_string());
 
         assert_eq!(
@@ -699,23 +664,21 @@ mod tests {
             .unwrapped()
             .map(Expr::Literal);
 
-        let operator = choice((
-            left_infix(just('+'), 1, |l, r| Expr::Add(Box::new(l), Box::new(r))),
-            left_infix(just('-'), 1, |l, r| Expr::Sub(Box::new(l), Box::new(r))),
-            right_infix(just('*'), 2, |l, r| Expr::Mul(Box::new(l), Box::new(r))),
-            right_infix(just('/'), 2, |l, r| Expr::Div(Box::new(l), Box::new(r))),
-        ));
-
         let parser = atom
-            .pratt(operator)
-            .with_postfix_ops(choice((
-                // Because we defined '+' and '-' as left associative operators,
+            .pratt((
+                // -- Postfix
+                // Because we defined '*' and '/' as right associative operators,
                 // in order to get these to function as expected, their strength
-                // must be higher, i.e. they must bind tighter
-                postfix(just('!'), 2, |lhs| Expr::Factorial(Box::new(lhs))),
-                // Or weirdness happens
-                postfix(just('$'), 0, |lhs| Expr::Value(Box::new(lhs))),
-            )))
+                // must be higher
+                postfix(2, just('!'), |l| u(Expr::Factorial, l)),
+                // This is what happens when not
+                postfix(0, just('$'), |l| u(Expr::Value, l)),
+                // -- Infix
+                infix(left(1), just('+'), |l, r| i(Expr::Add, l, r)),
+                infix(left(1), just('-'), |l, r| i(Expr::Sub, l, r)),
+                infix(right(2), just('*'), |l, r| i(Expr::Mul, l, r)),
+                infix(right(2), just('/'), |l, r| i(Expr::Div, l, r)),
+            ))
             .map(|x| x.to_string());
 
         assert_eq!(
@@ -731,34 +694,22 @@ mod tests {
             .unwrapped()
             .map(Expr::Literal);
 
-        let operator = choice((
-            left_infix(just('+'), 1, |l, r| Expr::Add(Box::new(l), Box::new(r))),
-            left_infix(just('-'), 1, |l, r| Expr::Sub(Box::new(l), Box::new(r))),
-            right_infix(just('*'), 2, |l, r| Expr::Mul(Box::new(l), Box::new(r))),
-            right_infix(just('/'), 2, |l, r| Expr::Div(Box::new(l), Box::new(r))),
-        ));
-
         let parser = atom
-            .pratt(operator)
-            .with_prefix_ops(choice((
-                // Because we defined '*' and '/' as right associative operators,
-                // in order to get these to function as expected, their strength
-                // must be higher
-                prefix(just('-'), 4, |rhs| Expr::Negate(Box::new(rhs))),
-                prefix(just('~'), 4, |rhs| Expr::Not(Box::new(rhs))),
-                // This is what happens when not
-                prefix(just('§'), 1, |rhs| Expr::Confusion(Box::new(rhs))),
-            )))
-            .with_postfix_ops(choice((
-                // Because we defined '+' and '-' as left associative operators,
-                // in order to get these to function as expected, their strength
-                // must be higher, i.e. they must bind tighter
-                postfix(just('!'), 5, |lhs| Expr::Factorial(Box::new(lhs))),
-                // Or weirdness happens
-                postfix(just('$'), 0, |lhs| Expr::Value(Box::new(lhs))),
-            )))
+            .pratt((
+                // -- Prefix
+                prefix(4, just('-'), |r| u(Expr::Negate, r)),
+                prefix(4, just('~'), |r| u(Expr::Not, r)),
+                prefix(1, just('§'), |r| u(Expr::Confusion, r)),
+                // -- Postfix
+                postfix(5, just('!'), |l| u(Expr::Factorial, l)),
+                postfix(0, just('$'), |l| u(Expr::Value, l)),
+                // -- Infix
+                infix(left(1), just('+'), |l, r| i(Expr::Add, l, r)),
+                infix(left(1), just('-'), |l, r| i(Expr::Sub, l, r)),
+                infix(right(2), just('*'), |l, r| i(Expr::Mul, l, r)),
+                infix(right(2), just('/'), |l, r| i(Expr::Div, l, r)),
+            ))
             .map(|x| x.to_string());
-
         assert_eq!(
             parser.parse("§1+-~2!$*3").into_result(),
             Ok("(((§(1 + (-(~(2!)))))$) * 3)".to_string()),
